allow user to choose camera

+++ app.js
@@ -32,7 +32,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 /* GET home page. */
 var router = express.Router();
 router.get('/', function(req, res) {
-    res.render('index', { title: 'Express' });
+    res.render('index', { title: 'enNodr' });
 });
 app.use('/', router);
 
@@ -68,19 +68,23 @@ app.use(function(err, req, res, next) {
     });
 });
 
+var i = 1;
 io.on('connection', function(socket){
-    var encChild = spawn('avconv', "-y -f image2pipe -s 1280x720 -c:v mjpeg -pix_fmt rgb24 -i - -an -c:v libx264 -pix_fmt yuv420p -f mpegts out/test.ts".split(" "));
+    var encChild;
 
-    encChild.on('close', function(code){
-        console.log('avconv closed: ' + code);
-    })
-    encChild.stderr.setEncoding('utf8')
-    encChild.stderr.on('data', function(data){
-        console.log(data);
-    });
-    encChild.stdout.setEncoding('utf8')
-    encChild.stdout.on('data', function(data){
-        console.log(data);
+    socket.once('frame', function(){
+        encChild = spawn('avconv', ("-y -f image2pipe -c:v mjpeg -pix_fmt rgb24 -r 30 -i - -ar 48000 -ac 2 -f s16le -i /dev/zero -shortest -c:v libx264 -c:a aac -strict experimental -pix_fmt yuv420p -f mpegts out/test" + (i++) + ".ts").split(" "));
+        encChild.stdin.on('error', function(e){ console.error(e); });
+        encChild.stderr.on('error', function(e){ console.error(e); });
+
+        encChild.on('close', function(code){
+            console.log('avconv closed: ' + code);
+        });
+
+        encChild.stderr.setEncoding('utf8')
+        encChild.stderr.on('data', function(data){
+            console.log(data);
+        });
     });
     
     socket.on('frame', function(b64){
@@ -89,8 +93,9 @@ io.on('connection', function(socket){
     });
 
     socket.on('disconnect', function(){
-        encChild.kill();
-    })
+        socket.removeAllListeners('frame');
+        if (encChild) encChild.kill();
+    });
 });
